Remove: Delete InputView Feat: Add PokerGame
